refactor(Job): store path to state root
Rename `config_dir` to `config_root` and `state_dir` to `state_root`
in `Config`.
While working on the rest of this refactor,
I confused config and job `state_dir` and created a bug
(fixed now).

I have avoided duplicating global configuration in jobs.
In the case of `state_root`,
this has made the interface of the class `Job` worse
(`state_dir` arguments everywhere),
prevented caching,
and led to more repeative code outside the class,
a potential source of bugs.
Moreover, there is no practical need to support
dynamic changes to global config now.
If a need arises,
the program can reload the jobs.